requirements.txt
langchain-astradb
langchain 
langchain-openai
datasets
pypdf
python-dotenv
flask
langchain_core
langchain_google_genai
-e .